+++ lib/components/forms/UluFormActions.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="form-theme__actions">
+    <slot />
+  </div>
+</template>
+
+<script setup>
+// This component doesn't need any script.
+</script>