+++ lib/components/forms/UluFormRequiredChar.vue
@@ -1,0 +1,7 @@
+<template>
+  <span class="form-theme__required-char">*</span>
+</template>
+
+<script setup>
+// This component doesn't need any script.
+</script>